restructure build a bit
- move generator code to core
- make root project an aggregate of core, svc, and web
- remove duplicate of core's UrlKey object from web
