+++ web/app/controllers/codegen/RubyGenerator.scala
@@ -5,7 +5,7 @@ import controllers.Authenticated
 import client.Apidoc.Version
 import play.api._
 import play.api.mvc._
-import lib.RubyGemGenerator
+import core.generator.RubyGemGenerator
 
 import play.api._
 import play.api.mvc._
