--- web/app/lib/UrlKey.scala
@@ -1,16 +1,0 @@
-package lib
-
-object UrlKey {
-
-  // Only want lower case letters, dashes and period
-  private val Regexp1 = """([^0-9a-z\-\.])""".r
-
-  // Now turn multiple dashes into single dashes
-  private val Regexp2 = """(\-+)""".r
-
-  def generate(value: String): String = {
-    val a = Regexp1.replaceAllIn(value.toLowerCase.trim, m => "-")
-    Regexp2.replaceAllIn(a, m => "-")
-  }
-
-}
